# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        description: Error message
        type: string
    type: object
  models.MyTraining:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: integer
      num_places:
        type: integer
      status:
        type: string
      trainer_id:
        type: integer
      training_date:
        type: string
      training_time:
        type: string
      updated_at:
        type: string
    type: object
...
  models.TrainingGroup:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      trainer_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.TrainingRegistration:
    properties:
      id:
        type: integer
      registration_date:
        type: string
      status:
        type: string
      training_session_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.TrainingSession:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      num_places:
        type: integer
      start_time:
        type: string
      trainer_id:
        type: integer
      training_date:
        type: string
      training_type:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password_hash:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      wallet_balance:
        type: number
    type: object
host: 25.28.121.106:8080
info:
  contact: {}
  description: API для Basketball Academy
  title: Basketball Academy API
  version: "1.0"
paths:
...
  /api/TrainingGroup:
    get:
      consumes:
      - application/json
      description: Get a list of all TrainingGroup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TrainingGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all TrainingGroup
      tags:
      - TrainingGroup
    post:
      consumes:
      - application/json
      description: Create a new TrainingGroup with the given information
      parameters:
      - description: TrainingGroup data
        in: body
        name: TrainingGroup
        required: true
        schema:
          $ref: '#/definitions/models.TrainingGroup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TrainingGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new TrainingGroup
      tags:
      - TrainingGroup
...
  /api/my-trainings/{id}/cancel:
    post:
      parameters:
      - description: ID записи в MyTraining
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MyTraining'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отменить запись на тренировку (в таблице MyTraining)
      tags:
      - Trainings
...
  /api/training-registrations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные регистрации на тренировку
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.TrainingRegistration'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TrainingRegistration'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Записаться на тренировку
      tags:
      - Trainings
  /api/training-sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TrainingSession'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить расписание тренировок
      tags:
      - Trainings
...
  /api/training-sessions/{sessionId}/participants:
    get:
      parameters:
      - description: ID тренировки
        in: path
        name: sessionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TrainingRegistration'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить участников тренировки по ID сессии
      tags:
      - Trainings
  /api/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the given information
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user data by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update user
      tags:
      - users
swagger: "2.0"